backend/requirements.txt
fastapi
uvicorn
firebase-admin
numpy
hypothesis
pytest
python-dotenv==1.0.1
